My Projects: Get projects from user context rather than flux call
