+++ common/components/controllers/MyProjectsController.jsx
@@ -4,13 +4,7 @@ import CurrentUser from "../utils/CurrentUser.js";
 import ProjectAPIUtils from "../utils/ProjectAPIUtils.js";
 import MyProjectCard from "../componentsBySection/MyProjects/MyProjectCard.jsx";
 import ConfirmationModal from "../common/confirmation/ConfirmationModal.jsx";
-import MyProjectsStore, {
-  MyProjectData,
-  MyProjectsAPIResponse,
-} from "../stores/MyProjectsStore.js";
-import UniversalDispatcher from "../stores/UniversalDispatcher.js";
 import metrics from "../utils/metrics.js";
-import { Container } from "flux/utils";
 import ProjectVolunteerRenewModal from "../common/projects/ProjectVolunteerRenewModal.jsx";
 import ProjectVolunteerConcludeModal from "../common/projects/ProjectVolunteerConcludeModal.jsx";
 import LogInController from "./LogInController.jsx";
@@ -19,6 +13,7 @@ import Section from "../enums/Section";
 import React from "react";
 import _ from "lodash";
 import Headers from "../common/Headers.jsx";
+import type { UserContext } from "../utils/CurrentUser.js";
 
 type State = {|
   ownedProjects: ?Array<MyProjectData>,
@@ -28,35 +23,20 @@ type State = {|
   showConcludeVolunteerModal: boolean,
 |};
 
-class MyProjectsController extends React.Component<{||}, State> {
+class MyProjectsController extends React.PureComponent<{||}, State> {
   constructor(): void {
     super();
+    const userContext: UserContext = CurrentUser.userContext();
     this.state = {
-      ownedProjects: null,
-      volunteeringProjects: null,
+      ownedProjects: userContext.owned_projects,
+      volunteeringProjects: userContext.volunteering_projects,
       showConfirmDeleteModal: false,
       showRenewVolunteerModal: false,
       showConcludeVolunteerModal: false,
     };
   }
 
-  static getStores(): $ReadOnlyArray<FluxReduceStore> {
-    return [MyProjectsStore];
-  }
-
-  static calculateState(prevState: State): State {
-    const myProjects: MyProjectsAPIResponse = MyProjectsStore.getMyProjects();
-    return {
-      ownedProjects: myProjects && myProjects.owned_projects,
-      volunteeringProjects: myProjects && myProjects.volunteering_projects,
-    };
-  }
-
   componentWillMount(): void {
-    setTimeout(function() {
-      // Run after dispatcher has finished
-      UniversalDispatcher.dispatch({ type: "INIT" });
-    }, 0);
     const args = url.arguments(window.location.href);
     if (
       "from" in args &&
@@ -152,6 +132,7 @@ class MyProjectsController extends React.Component<{||}, State> {
   }
 
   render(): React$Node {
+    // TODO: Move header to backend
     return CurrentUser.isLoggedIn() ? (
       <React.Fragment>
         <Headers
@@ -221,4 +202,4 @@ class MyProjectsController extends React.Component<{||}, State> {
   }
 }
 
-export default Container.create(MyProjectsController);
+export default MyProjectsController;
