+++ common/components/stores/UniversalDispatcher.js
@@ -2,7 +2,6 @@
 
 "use strict";
 
-import type { MyProjectsActionType } from "./MyProjectsStore.js";
 import type { MyGroupsActionType } from "./MyGroupsStore.js";
 import type { NavigationActionType } from "./NavigationStore.js";
 import type { OffsetActionType } from "./PageOffsetStore.js";
@@ -11,7 +10,6 @@ import type { FormFieldsActionType } from "./FormFieldsStore.js";
 import { Dispatcher } from "flux";
 
 type UniversalActionType =
-  | MyProjectsActionType
   | MyGroupsActionType
   | NavigationActionType
   | OffsetActionType
